pyproject.toml
[tool.poetry]
name = "aliceBot"
version = "0.1.0"
description = "aliceBot"
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7.3"
nonebot2 = "^2.0.0.a1"

[tool.poetry.dev-dependencies]
nb-cli = "^0.5.0"
nonebot-plugin-test = "^0.3.0"

[nonebot.plugins]
plugins = ["nonebot_plugin_apscheduler"]
plugin_dirs = ["alicebot/plugins"]

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
